extract duplicate name validation tests to its own file, add test for full name mapper via client scope
