+++ provider/keycloak_openid_full_name_protocol_mapper_test.go
@@ -212,31 +212,6 @@ func TestAccKeycloakOpenIdFullNameProtocolMapper_updateClientScopeForceNew(t *te
 	})
 }
 
-/* Keycloak does not allow two protocol mappers to exist with the same name, but if you create two of them with the same
- * name quickly enough (which Terraform is pretty good at), the API allows it.  So this test needs to create a different
- * mapper first, then attempt to create the full name mapper with the same name in order to get the expected error
- */
-func TestAccKeycloakOpenIdFullNameProtocolMapper_clientDuplicateNameValidation(t *testing.T) {
-	realmName := "terraform-realm-" + acctest.RandString(10)
-	clientId := "terraform-client-" + acctest.RandString(10)
-	mapperName := "terraform-openid-connect-full-name-mapper-" + acctest.RandString(5)
-
-	resource.Test(t, resource.TestCase{
-		Providers:    testAccProviders,
-		PreCheck:     func() { testAccPreCheck(t) },
-		CheckDestroy: testAccKeycloakOpenIdFullNameProtocolMapperDestroy(),
-		Steps: []resource.TestStep{
-			{
-				Config: testKeycloakOpenIdFullNameProtocolMapper_clientDuplicateNameValidationWithGroupMapper(realmName, clientId, mapperName),
-			},
-			{
-				Config:      testKeycloakOpenIdFullNameProtocolMapper_clientDuplicateNameValidationWithGroupAndFullNameMapper(realmName, clientId, mapperName),
-				ExpectError: regexp.MustCompile("validation error: a protocol mapper with name .+ already exists for this client"),
-			},
-		},
-	})
-}
-
 func testAccKeycloakOpenIdFullNameProtocolMapperDestroy() resource.TestCheckFunc {
 	return func(state *terraform.State) error {
 		for resourceName, rs := range state.RootModule().Resources {
@@ -421,54 +396,4 @@ resource "keycloak_openid_full_name_protocol_mapper" "full_name_mapper_validatio
 	name       = "%s"
 	realm_id   = "${keycloak_realm.realm.id}"
 }`, realmName, mapperName)
-}
-
-func testKeycloakOpenIdFullNameProtocolMapper_clientDuplicateNameValidationWithGroupMapper(realmName, clientId, mapperName string) string {
-	return fmt.Sprintf(`
-resource "keycloak_realm" "realm" {
-	realm = "%s"
-}
-
-resource "keycloak_openid_client" "openid_client" {
-	realm_id    = "${keycloak_realm.realm.id}"
-	client_id   = "%s"
-
-	access_type = "BEARER-ONLY"
-}
-
-resource "keycloak_openid_group_membership_protocol_mapper" "group_membership_mapper_client" {
-  	name            = "%s"
-	realm_id        = "${keycloak_realm.realm.id}"
-  	client_id       = "${keycloak_openid_client.openid_client.id}"
-
-  	claim_name      = "foo"
-}`, realmName, clientId, mapperName)
-}
-
-func testKeycloakOpenIdFullNameProtocolMapper_clientDuplicateNameValidationWithGroupAndFullNameMapper(realmName, clientId, mapperName string) string {
-	return fmt.Sprintf(`
-resource "keycloak_realm" "realm" {
-	realm = "%s"
-}
-
-resource "keycloak_openid_client" "openid_client" {
-	realm_id    = "${keycloak_realm.realm.id}"
-	client_id   = "%s"
-
-	access_type = "BEARER-ONLY"
-}
-
-resource "keycloak_openid_group_membership_protocol_mapper" "group_membership_mapper_client" {
-  	name            = "%s"
-	realm_id        = "${keycloak_realm.realm.id}"
-  	client_id       = "${keycloak_openid_client.openid_client.id}"
-
-  	claim_name      = "foo"
-}
-
-resource "keycloak_openid_full_name_protocol_mapper" "full_name_mapper_client" {
-  	name       = "%s"
-	realm_id   = "${keycloak_realm.realm.id}"
-  	client_id  = "${keycloak_openid_client.openid_client.id}"
-}`, realmName, clientId, mapperName, mapperName)
 }
